+++ BoxContentSDK/BoxContentSDK/Requests/BOXFileRequest.m
@@ -16,12 +16,7 @@
 
 @property (nonatomic, readwrite, strong) NSString *fileID;
 @property (nonatomic, readwrite, assign) BOOL isTrashed;
-
-/**
- Setting a representation or list option will include availability of file with request representation information
- */
-@property (nonatomic, readwrite, strong) NSMutableOrderedSet *representationsRequested;
-
+@property (nonatomic, readonly, strong) BOXRepresentationsHelper* repsHelper;
 
 - (BOOL)shouldPerformBackgroundOperation;
 
@@ -39,6 +34,7 @@
     if (self = [super init]) {
         _fileID = fileID;
         _isTrashed = isTrashed;
+        _repsHelper = [BOXRepresentationsHelper new];
     }
 
     return self;
@@ -65,16 +61,9 @@
         fieldString = [self fullFileFieldsParameterString];
     }
     
-    // The original content url is retrieved from the download_url. Include BOXRepresentationRequestOriginal
-    // The download url is added to the BoxFile representations array, retrieved by type value BOXRepresentationRequestOriginal.
-    // You can download the file has owner, co-owner, editor, viewer uploader, viewer permissions.
-    if ([self.representationsRequested containsObject:@(BOXRepresentationRequestOriginal)]) {
-        [self.representationsRequested removeObject:@(BOXRepresentationRequestOriginal)];
-        fieldString = [fieldString stringByAppendingFormat:@",%@", BOXAPIObjectKeyAuthenticatedDownloadURL];
-    }
-    if ([self.representationsRequested count] > 0) {
-        // Include information for the original content URL in any request for file representations
-        fieldString = [fieldString stringByAppendingFormat:@",%@", BOXAPIObjectKeyRepresentations];
+    NSString *repsFieldString = [self.repsHelper getRepresentationsFieldString];
+    if(repsFieldString.length > 0) {
+        fieldString = [fieldString stringByAppendingFormat:@",%@", repsFieldString];
     }
     
     if (fieldString.length > 0) {
@@ -113,7 +102,7 @@
         }
     }
     
-    NSString *representationFields = [self formatRepresentationRequestHeader];
+    NSString *representationFields = [self.repsHelper formatRepresentationRequestHeader];
     
     if (representationFields.length > 0) {
         [fileOperation.APIRequest addValue:representationFields
@@ -224,6 +213,10 @@
     [self performRequestWithCompletion:refreshBlock];
 }
 
+- (void)setRepresentationRequestOptions:(NSArray *)representationOptions {
+    [self.repsHelper setRepresentationRequestOptions:representationOptions];
+}
+
 #pragma mark - Superclass overidden methods
 
 - (NSString *)itemIDForSharedLink
@@ -243,77 +236,6 @@
     return (self.associateID.length > 0 && self.requestDirectoryPath.length > 0);
 }
 
-- (void)setRepresentationRequestOptions:(NSArray *)representationOptions
-{
-    self.representationsRequested = [[NSMutableOrderedSet alloc] initWithArray:representationOptions];
-}
 
-- (NSString *)formatRepresentationRequestHeader
-{
-    // Only process valid API options.
-    if ([self.representationsRequested containsObject:@(BOXRepresentationRequestOriginal)]) {
-        [self.representationsRequested removeObject:@(BOXRepresentationRequestOriginal)];
-    }
-    
-    if ([self.representationsRequested count] == 0) {
-        return nil;
-    }
-    
-    __block NSString *representationFields = @"";
-    
-    if ([self.representationsRequested containsObject:@(BOXRepresentationRequestAllRepresentations)]) {
-        representationFields = @"[jpg?dimensions=320x320&paged=false][jpg?dimensions=1024x1024&paged=false][pdf,hls,mp4,mp3,jpg]";
-        [self.representationsRequested removeObject:@(BOXRepresentationRequestAllRepresentations)];
-    }
-    if ([self.representationsRequested count] == 0) {
-        return representationFields;
-    }
-    
-    representationFields = [representationFields stringByAppendingString:@"["];
-    
-    __block NSString *delimiter = @"],[";
-    if (self.matchSupportedRepresentation == YES) {
-        delimiter = @",";
-    }
-    
-    [self.representationsRequested enumerateObjectsUsingBlock:^(id obj, NSUInteger idx, BOOL *stop) {
-        if (idx == [self.representationsRequested count]) {
-            representationFields = [representationFields stringByReplacingCharactersInRange:NSMakeRange([representationFields length]-1, 1) withString:@"]"];
-            
-            *stop = YES;
-        } else {
-            if (idx == [self.representationsRequested  count] - 1) {
-                delimiter = @"]";
-            }
-            BOXRepresentationRequestOptions representationOption = (BOXRepresentationRequestOptions) [obj integerValue];
-            if (representationOption == BOXRepresentationRequestHighDefinitionVideo) {
-                representationFields = [representationFields stringByAppendingString:[NSString stringWithFormat:@"%@%@", BOXRepresentationTypeHLS, delimiter]];
-            }
-            if (representationOption == BOXRepresentationRequestMP3Representation) {
-                representationFields = [representationFields stringByAppendingString:[NSString stringWithFormat:@"%@%@", BOXRepresentationTypeMP3, delimiter]];
-            }
-            if (representationOption == BOXRepresentationRequestMP4Representation) {
-                representationFields = [representationFields stringByAppendingString:[NSString stringWithFormat:@"%@%@", BOXRepresentationTypeMP4, delimiter]];
-            }
-            if (representationOption == BOXRepresentationRequestThumbnailRepresentation) {
-                representationFields = [representationFields stringByAppendingString:[NSString stringWithFormat:@"%@?dimensions=%@&paged=false%@", BOXRepresentationTypeJPG, BOXRepresentationImageDimensionsJPG320, delimiter]];
-            }
-            if (representationOption == BOXRepresentationRequestLargeThumbnailRepresentation) {
-                representationFields = [representationFields stringByAppendingString:[NSString stringWithFormat:@"%@?dimensions=%@&paged=false%@", BOXRepresentationTypeJPG, BOXRepresentationImageDimensions1024, delimiter]];
-            }
-            if (representationOption == BOXRepresentationRequestPDFRepresentation) {
-                representationFields = [representationFields stringByAppendingString:[NSString stringWithFormat:@"%@%@", BOXRepresentationTypePDF, delimiter]];
-            }
-            if (representationOption == BOXRepresentationRequestJPGRepresentation) {
-                representationFields = [representationFields stringByAppendingString:[NSString stringWithFormat:@"%@?dimensions=%@&paged=false%@", BOXRepresentationTypeJPG, BOXRepresentationImageDimensions1024, delimiter]];
-            }
-            if (representationOption == BOXRepresentationRequesteExtractedTextRepresentation) {
-                representationFields = [representationFields stringByAppendingString:[NSString stringWithFormat:@"%@%@", BOXRepresentationTypeExtractedText, delimiter]];
-            }
-        }
-    }];
-    
-    return representationFields;
-}
 
 @end
